Serve public quiz images as static files in the browser, private images via controller Add HasUnfinishedAttempt and PublicImageUrl to QuizDto Add route to change quiz privacy Introduce FilterByUnfinishedAttempts to get all quizzes

+++ QuizAI.Application/Services/FileStorageService.cs
@@ -1,55 +1,68 @@
-﻿using Microsoft.Extensions.Hosting;
-using QuizAI.Application.Interfaces;
+﻿using QuizAI.Application.Interfaces;
+using QuizAI.Domain.Enums;
 using QuizAI.Domain.Exceptions;
 
 namespace QuizAI.Application.Services;
 
 public class FileStorageService : IFileStorageService
 {
-    private readonly string _storagePath;
+    private readonly string _publicStoragePath;
+    private readonly string _privateStoragePath;
 
-    public FileStorageService(IHostEnvironment environment, string folderName)
+    public FileStorageService(string publicStoragePath, string privateStoragePath)
     {
-        _storagePath = Path.Combine(environment.ContentRootPath, folderName);
-
-        if (!Directory.Exists(_storagePath))
-        {
-            Directory.CreateDirectory(_storagePath);
-        }
+        _publicStoragePath = publicStoragePath;
+        _privateStoragePath = privateStoragePath;
     }
 
-    public async Task<Guid> UploadAsync(byte[] fileData, string fileExtension)
+    public async Task<Guid> UploadAsync(byte[] fileData, string fileExtension, bool isPrivate, Guid? filename = null)
     {
-        var filenameAsGuid = Guid.NewGuid();
-        var filePath = Path.Combine(_storagePath, filenameAsGuid + fileExtension.ToLower());
+        if (filename == null) filename = Guid.NewGuid();
+
+        var fullFilename = filename.ToString() + fileExtension.ToLower();
+        var filePath = Path.Combine(GetProperStoragePath(isPrivate), fullFilename);
+
+        if (File.Exists(filePath))
+            throw new ConflictException($"The file {fullFilename} already exists");
 
         await File.WriteAllBytesAsync(filePath, fileData);
 
-        return filenameAsGuid;
+        return (Guid)filename;
     }
 
-    public async Task<byte[]> RetrieveAsync(Guid filename, string fileExtension)
+    public async Task<byte[]> RetrieveAsync(Guid filename, string fileExtension, bool isPrivate)
     {
-        var filePath = GetFullFilePathIfExists(filename, fileExtension);
+        var filePath = GetFullFilePathIfExists(filename, fileExtension, isPrivate);
 
         return await File.ReadAllBytesAsync(filePath);
     }
 
-    public void Delete(Guid filename, string fileExtension)
+    public void Delete(Guid filename, string fileExtension, bool isPrivate)
     {
-        var filePath = GetFullFilePathIfExists(filename, fileExtension);
+        var filePath = GetFullFilePathIfExists(filename, fileExtension, isPrivate);
 
         File.Delete(filePath);
     }
 
-    private string GetFullFilePathIfExists(Guid filename, string fileExtension)
+    public void CopyImage(Guid filename, string fileExtension, bool isPrivate)
+    {
+        string oldFullFilePath = GetFullFilePathIfExists(filename, fileExtension, isPrivate);
+        string newFullFilePath = Path.Combine(GetProperStoragePath(!isPrivate), filename.ToString() + fileExtension.ToLower());
+
+        File.Copy(oldFullFilePath, newFullFilePath, overwrite: false);
+    }
+
+    private string GetFullFilePathIfExists(Guid filename, string fileExtension, bool isPrivate)
     {
         var fullFilename = filename.ToString() + fileExtension.ToLower();
-        var filePath = Path.Combine(_storagePath, fullFilename);
+        var filePath = Path.Combine(GetProperStoragePath(isPrivate), fullFilename);
 
         if (!File.Exists(filePath))
             throw new NotFoundException($"The file {fullFilename} could not be found");
 
         return filePath;
     }
+
+    private string GetProperStoragePath(bool isPrivate) => 
+        isPrivate ? _privateStoragePath : _publicStoragePath;
 }
